pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-reliability-engine"
version = "1.0.0"
description = "Production-grade, failure-aware RAG system with retrieval quality gating and answer verification"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # LLM providers
    "google-genai>=1.0.0",
    "openai>=1.40.0",

    # Vector store
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",

    # Keyword search
    "rank-bm25>=0.2.2",

    # Cross-encoder reranker
    "sentence-transformers>=3.0.0",

    # Document parsing
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.10",
    "beautifulsoup4>=4.12.0",
    "charset-normalizer>=3.0.0",

    # Tokenization
    "tiktoken>=0.7.0",

    # Configuration
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",

    # Database
    "aiosqlite>=0.20.0",

    # Observability
    "structlog>=24.0.0",

    # Near-duplicate detection
    "datasketch>=1.6.0",

    # Language detection
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rag_engine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
